refactor: 확장함수 AppContainer 내부로 이동
