+++ hashdi/src/main/java/woowacourse/shopping/hashdi/Injector.kt
@@ -30,7 +30,7 @@ class Injector(private val appContainer: AppContainer) {
     private fun <T : Any> fieldInjection(clazz: KClass<T>, target: T) {
         clazz.declaredMemberProperties.forEach { property ->
             if (property.hasAnnotation<Inject>()) {
-                val injectValue = appContainer.getInstance(property.identifyKey())
+                val injectValue = appContainer.getInstance(property)
                 property.isAccessible = true
                 (property as KMutableProperty<*>).setter.call(target, injectValue)
             }
@@ -39,7 +39,7 @@ class Injector(private val appContainer: AppContainer) {
 
     private fun getArgumentsMapping(parameters: List<KParameter>): Map<KParameter, Any> {
         return parameters.associateWith { param ->
-            appContainer.getInstance(param.identifyKey())
+            appContainer.getInstance(param)
         }
     }
 }
